+++ lib/site-inspector.rb
@@ -8,11 +8,11 @@ require 'typhoeus'
 require_relative 'site-inspector/cache'
 require_relative 'site-inspector/compliance'
 require_relative 'site-inspector/disk_cache'
-require_relative 'site-inspector/dns'
 require_relative 'site-inspector/domain'
 require_relative 'site-inspector/endpoint'
-require_relative 'site-inspector/headers'
-require_relative 'site-inspector/hsts'
+require_relative 'site-inspector/endpoint/dns'
+require_relative 'site-inspector/endpoint/headers'
+require_relative 'site-inspector/endpoint/hsts'
 require_relative 'site-inspector/sniffer'
 require_relative 'site-inspector/version'
