+++ lib/site-inspector/endpoint/dns.rb
