+++ lib/site-inspector/endpoint/headers.rb
